+++ src/main/resources/mapper/ScheduleMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.hrm.mapper.ScheduleMapper">
+<mapper namespace="com.hrm.dao.ScheduleDao">
 
     <!-- ✅ 모든 일정 조회 -->
     <select id="getAllSchedules" resultType="com.hrm.dto.ScheduleDto">
